Move source_name back to kafka.helpers
